Remove arrow.py
